pyproject.toml
[project]
name = "bertalign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "numpy<2",
    "googletrans>=4.0.2",
    "numba>=0.62.1",
    "sentence-splitter>=1.4",
    "sentence-transformers>=5.1.2",
    "faiss-gpu>=1.7.2",
    "psutil>=7.1.3",
    "langdetect>=1.0.9",
    "openai>=1.0.0",
]
